CSharp Workspace (WIP)

+++ Core/Workspace/CSharp/Adapters/Allors.Workspace.Adapters.Local/Session/OriginState/SessionOriginState.cs
@@ -19,11 +19,14 @@ namespace Allors.Workspace.Adapters.Local
             this.propertyByObjectByPropertyType = new PropertyByObjectByPropertyType(numbers);
         }
 
-        internal void Checkpoint(ChangeSet changeSet) => changeSet.AddSessionStateChanges(this.propertyByObjectByPropertyType.Checkpoint());
+        internal void Checkpoint(ChangeSet changeSet) =>
+            changeSet.AddSessionStateChanges(this.propertyByObjectByPropertyType.Checkpoint());
 
-        internal object Get(long @object, IPropertyType propertyType) => this.propertyByObjectByPropertyType.Get(@object, propertyType);
+        internal object Get(long @object, IPropertyType propertyType) =>
+            this.propertyByObjectByPropertyType.Get(@object, propertyType);
 
-        internal void SetUnitRole(long association, IRoleType roleType, object role) => this.propertyByObjectByPropertyType.Set(association, roleType, role);
+        internal void SetUnitRole(long association, IRoleType roleType, object role) =>
+            this.propertyByObjectByPropertyType.Set(association, roleType, role);
 
         internal void SetCompositeRole(long association, IRoleType roleType, long? newRole)
         {
@@ -45,7 +48,8 @@ namespace Allors.Workspace.Adapters.Local
             if (associationType.IsOne)
             {
                 // OneToOne
-                var previousAssociation = (long?)this.propertyByObjectByPropertyType.Get(newRole.Value, associationType);
+                var previousAssociation =
+                    (long?)this.propertyByObjectByPropertyType.Get(newRole.Value, associationType);
                 if (previousAssociation.HasValue)
                 {
                     this.propertyByObjectByPropertyType.Set(previousAssociation.Value, roleType, null);
@@ -97,12 +101,15 @@ namespace Allors.Workspace.Adapters.Local
             // Association
             if (associationType.IsOne)
             {
-                var previousRoleAssociations = this.propertyByObjectByPropertyType.Get((long)previousRole, associationType);
-                this.propertyByObjectByPropertyType.Set((long)previousRole, associationType, this.numbers.Remove(previousRoleAssociations, association));
+                var previousRoleAssociations =
+                    this.propertyByObjectByPropertyType.Get((long)previousRole, associationType);
+                this.propertyByObjectByPropertyType.Set((long)previousRole, associationType,
+                    this.numbers.Remove(previousRoleAssociations, association));
             }
 
-            var roleAssociations = this.propertyByObjectByPropertyType.Get((long)roleToAdd, associationType);
-            this.propertyByObjectByPropertyType.Set(roleToAdd, associationType, this.numbers.Add(roleAssociations, association));
+            var roleAssociations = this.propertyByObjectByPropertyType.Get(roleToAdd, associationType);
+            this.propertyByObjectByPropertyType.Set(roleToAdd, associationType,
+                this.numbers.Add(roleAssociations, association));
         }
 
         private void RemoveRole(long association, IRoleType roleType, long roleToRemove)
